checkpoint, creating valuation agent fix
