+++ server/src/api/controllers/analysis.py
@@ -15,18 +15,12 @@ logger = logging.getLogger(__name__)
 def generate_stream(request_data: AnalysisRequest) -> Generator[str, None, None]:
     """Generate a stream of analysis data and progress updates."""
     try:
-        logger.debug("Starting stream generation")
-        # Create a new workflow for this request
-        workflow = create_workflow()
-        logger.debug("Workflow created")
-        
-        # Process the request and stream the results
         logger.debug("Starting to process analysis request")
         for result in process_analysis_request(request_data):
             if result is None:
                 logger.warning("Received None result from process_analysis_request")
                 continue
-                
+
             # Add timestamp to each event
             event = {
                 **result,
@@ -34,7 +28,7 @@ def generate_stream(request_data: AnalysisRequest) -> Generator[str, None, None]
             }
             # Send as raw JSON
             yield json.dumps(event) + "\n"
-            
+
     except Exception as e:
         logger.error(f"Error in stream generation: {str(e)}", exc_info=True)
         error_event = {
@@ -44,32 +38,9 @@ def generate_stream(request_data: AnalysisRequest) -> Generator[str, None, None]
         }
         yield json.dumps(error_event) + "\n"
 
-def handle_analysis_request(request_data: AnalysisRequest = None) -> Response:
+def handle_analysis_request(request_data: AnalysisRequest) -> Response:
     """Handle the analysis request and return a streaming response."""
     try:
-        logger.debug("Handling analysis request")
-        if request_data is None:
-            data = request.get_json()
-            if 'portfolio' in data and isinstance(data['portfolio'], dict):
-                portfolio_data = data['portfolio']
-                if 'positions' in portfolio_data:
-                    portfolio_data['positions'] = {
-                        ticker: Position(**pos_data) 
-                        for ticker, pos_data in portfolio_data['positions'].items()
-                    }
-                if 'realized_gains' in portfolio_data:
-                    portfolio_data['realized_gains'] = {
-                        ticker: RealizedGains(**gains_data)
-                        for ticker, gains_data in portfolio_data['realized_gains'].items()
-                    }
-                data['portfolio'] = Portfolio(**portfolio_data)
-            try:
-                request_data = AnalysisRequest(**data)
-                logger.debug("Successfully created AnalysisRequest")
-            except TypeError as e:
-                logger.error(f"Invalid request parameters: {str(e)}")
-                return jsonify({'error': f'Invalid request parameters: {str(e)}'}), 400
-        
         def stream():
             logger.debug("Starting stream")
             for chunk in generate_stream(request_data):
@@ -87,7 +58,7 @@ def handle_analysis_request(request_data: AnalysisRequest = None) -> Response:
                 'Access-Control-Allow-Origin': '*'
             }
         )
-        
+
     except Exception as e:
         logger.error(f"Error in request handling: {str(e)}", exc_info=True)
-        return jsonify({'error': str(e)}), 500 
+        return jsonify({'error': str(e)}), 500
